+++ love/src/components/AuxTel/Dome/Dome.container.jsx
@@ -99,11 +99,11 @@ const DomeContainer = ({
   subscribeToStream,
   unsubscribeToStream,
   controls,
-  atDomeTracking,
   targetName,
-  telescopeRA,
-  telescopeDec,
-  telescopeRotator,
+  telescopeRAHour,
+  telescopeRADeg,
+  telescopeDecDeg,
+  telescopeRotatorRad,
   raDecHourFormat,
   ...props
 }) => {
@@ -154,11 +154,11 @@ const DomeContainer = ({
       controls={controls}
       atDomeSummaryState={atDomeSummaryState}
       ATMCSSummaryState={ATMCSSummaryState}
-      atDomeTracking={atDomeTracking}
       targetName={targetName}
-      telescopeRA={telescopeRA}
-      telescopeDec={telescopeDec}
-      telescopeRotator={telescopeRotator}
+      telescopeRAHour={telescopeRAHour}
+      telescopeRADeg={telescopeRADeg}
+      telescopeDecDeg={telescopeDecDeg}
+      telescopeRotatorRad={telescopeRotatorRad}
       raDecHourFormat={raDecHourFormat}
     />
   );
@@ -177,6 +177,8 @@ const mapDispatchToProps = (dispatch) => {
     'telemetry-ATMCS-0-mount_AzEl_Encoders',
     'telemetry-ATMCS-0-mount_Nasmyth_Encoders',
     'telemetry-Scheduler-2-observatoryState',
+    'telemetry-ATPtg-0-mountStatus',
+    'telemetry-ATPtg-0-mountPosition',
     'event-ATDome-0-azimuthState',
     'event-ATDome-0-azimuthCommandedState',
     'event-ATDome-0-dropoutDoorState',
@@ -187,10 +189,10 @@ const mapDispatchToProps = (dispatch) => {
     'event-ATMCS-0-allAxesInPosition',
     'event-ATMCS-0-m3State',
     'event-ATMCS-0-positionLimits',
-    'event-ATPtg-1-timesOfLimits',
     'event-ATDome-0-summaryState',
     'event-ATMCS-0-summaryState',
-    `event-ATPtg-1-currentTarget`,
+    'event-ATPtg-0-timesOfLimits',
+    `event-ATPtg-0-currentTarget`,
   ];
   return {
     subscriptions,