+++ love/src/components/GeneralPurpose/ProgressBar/ProgressBar.jsx
@@ -18,56 +18,12 @@ this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
 import React, { useRef } from 'react';
+import PropTypes from 'prop-types';
 import styles from './ProgressBar.module.css';
 
-const ProgressBar = (props) => {
+const ProgressBar = ({ targetValue, completed, hideCompleted = false, height = 20 }) => {
   const ref = useRef(null);
-  const { targetValue, completed, hideCompleted = false, height = 20 } = props;
   const padding = 4;
-
-  const parentDiv = {
-    display: 'flex',
-  };
-
-  const containerStyles = {
-    height: 20,
-    width: '100%',
-    backgroundColor: '#455a64',
-    marginBottom: '5px',
-    float: 'rigth',
-    position: 'relative',
-    padding: `${padding}px`,
-  };
-
-  const fillerStyles = {
-    height: '100%',
-    width: `${completed}%`,
-    backgroundColor: 'var(--second-tertiary-background-color)',
-    textAlign: 'right',
-  };
-
-  const labelStyles = {
-    padding: 5,
-    color: 'white',
-    fontWeight: 'bold',
-  };
-
-  const targetValueLine = {
-    stroke: 'white',
-    strokeWidth: 2,
-    strokeDasharray: 1.8,
-    strokeOpacity: 0.5,
-  };
-
-  const progressCommandedLine = {
-    position: 'absolute',
-    top: padding,
-    right: padding,
-    bottom: padding,
-    left: padding,
-    overflow: 'visible',
-  };
-
   const width = ref.current?.clientWidth ?? 0;
   const targetValuePixels = ((width - 2 * padding) * targetValue) / 100;
   const completedValue = completed.toString().padStart(3, '0');
@@ -99,6 +55,17 @@ const ProgressBar = (props) => {
       </div>
     </div>
   );
+};
+
+ProgressBar.propTypes = {
+  /** Target value */
+  targetValue: PropTypes.number,
+  /** Completed value */
+  completed: PropTypes.number,
+  /** Hide completed value */
+  hideCompleted: PropTypes.bool,
+  /** Height of the progress bar */
+  height: PropTypes.number,
 };
 
 export default ProgressBar;