+++ love/src/components/MainTel/MTDome/MTDome.container.jsx
@@ -27,6 +27,7 @@ import {
   getDomeAzimuth,
   getLightWindScreen,
   getPointingStatus,
+  getMainTelescopeState,
 } from '../../../redux/selectors';
 import { addGroup, removeGroup } from '../../../redux/actions/ws';
 import SubscriptionTableContainer from '../../GeneralPurpose/SubscriptionTable/SubscriptionTable.container';
@@ -46,6 +47,12 @@ export const schema = {
       description: "Whether to display controls to configure periods of time'",
       default: true,
       isPrivate: false,
+    },
+    raDecHourFormat: {
+      type: 'boolean',
+      description: 'Whether to display the RA and DEC in hour format',
+      isPrivate: false,
+      default: false,
     },
   },
 };
@@ -72,7 +79,12 @@ const MTDomeContainer = ({
   targetPointingAz,
   currentPointingEl,
   targetPointingEl,
-  mtDomeTracking,
+  targetName,
+  telescopeRAHour,
+  telescopeRADeg,
+  telescopeDecDeg,
+  telescopeRotatorRad,
+  raDecHourFormat,
   ...props
 }) => {
   if (props.isRaw) {
@@ -101,7 +113,12 @@ const MTDomeContainer = ({
       targetPointingAz={targetPointingAz}
       currentPointingEl={currentPointingEl}
       targetPointingEl={targetPointingEl}
-      mtDomeTracking={mtDomeTracking}
+      targetName={targetName}
+      telescopeRAHour={telescopeRAHour}
+      telescopeRADeg={telescopeRADeg}
+      telescopeDecDeg={telescopeDecDeg}
+      telescopeRotatorRad={telescopeRotatorRad}
+      raDecHourFormat={raDecHourFormat}
     />
   );
 };
@@ -113,6 +130,7 @@ const mapStateToProps = (state) => {
   const lightWindScreenState = getLightWindScreen(state);
   const domeAzimuthState = getDomeAzimuth(state);
   const pointingState = getPointingStatus(state);
+  const telescopeState = getMainTelescopeState(state);
   return {
     ...domeState,
     ...louversState,
@@ -120,6 +138,7 @@ const mapStateToProps = (state) => {
     ...lightWindScreenState,
     ...domeAzimuthState,
     ...pointingState,
+    ...telescopeState,
   };
 };
 
@@ -131,6 +150,8 @@ const mapDispatchToProps = (dispatch) => {
     'telemetry-MTDome-0-louvers',
     'telemetry-MTMount-0-azimuth',
     'telemetry-MTMount-0-elevation',
+    'telemetry-MTPtg-0-mountStatus',
+    'telemetry-MTPtg-0-mountPosition',
     'event-MTDome-0-azEnabled',
     'event-MTDome-0-azMotion',
     'event-MTDome-0-azTarget',
@@ -138,6 +159,7 @@ const mapDispatchToProps = (dispatch) => {
     'event-MTMount-0-target',
     'event-MTDome-0-summaryState',
     'event-MTMount-0-summaryState',
+    'event-MTPtg-0-currentTarget',
   ];
   return {
     subscriptions,