+++ love/src/components/OLE/OLE.jsx
@@ -174,9 +174,8 @@ export default class OLE extends Component {
           registryMap[instrument] = key;
         });
       });
-      const instrumentsArray = Object.values(data)
-        .map((arr) => arr[0])
-        .filter((instrument) => instrument);
+
+      const instrumentsArray = Object.keys(registryMap);
 
       this.setState({
         instruments: instrumentsArray,